Add quantile encoding
